upda7-8

+++ BAL/Model/PurchaseOrderDetailModel.cs
@@ -8,7 +8,7 @@ namespace BAL.Model
     {
         public int PurchaseOrderDetailID { get; set; }
 
-        public int StockCount { get; set; }
+        public float StockCount { get; set; }
 
         
         public decimal PurchasePrice { get; set; }
